Move CategoryDropdown to root components
