--- src/pages/tokenLending/LendTokenPage/PlaceTokenOffers/components/FilterSection/constants.ts
@@ -1,13 +1,0 @@
-import { MarketCategory } from '@banx/api/tokens'
-
-export const MARKETS_CATEGORIES = [
-  { key: MarketCategory.All, label: 'All' },
-  { key: MarketCategory.LST, label: 'LST' },
-  { key: MarketCategory.DeFi, label: 'DeFi' },
-  { key: MarketCategory.Meme, label: 'Memes' },
-  { key: MarketCategory.Governance, label: 'Governance' },
-  { key: MarketCategory.RWA, label: 'RWA' },
-  { key: MarketCategory.LP, label: 'LP' },
-  { key: MarketCategory.DePin, label: 'DePin' },
-  { key: MarketCategory.Gaming, label: 'Gaming' },
-]
