+++ src/pages/tokenLending/OffersTokenPage/components/FilterSection/FilterSection.module.less
@@ -15,6 +15,13 @@
   max-width: 296px;
 }
 
+.filterContent {
+  display: flex;
+  align-items: center;
+  gap: 16px;
+  width: 100%;
+}
+
 .dropdownHidden {
   @media (max-width: 960px) {
     display: none;
